Added numeric selector

+++ InputDialog/NumericProperties.cs
@@ -2,10 +2,10 @@
 
 public class NumericProperties
 {
-    public int Value { get; set; } = 0;
-    public int Minimum { get; set; } = 0;
-    public int Maximum { get; set; } = 100;
-    public int Increment { get; set; } = 1;
+    public decimal Value { get; set; } = 0;
+    public decimal Minimum { get; set; } = 0;
+    public decimal Maximum { get; set; } = 100;
+    public decimal Increment { get; set; } = 1;
     public int DecimalPlaces { get; set; } = 0;
     public bool ThousandsSeparator { get; set; } = false;
     public HorizontalAlignment HorizontalAlignment { get; set; } = HorizontalAlignment.Right;
